+++ app/routes/privateRoutes/restClientPage/components/bodyEditor/BodyEditor.module.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   gap: var(--spacing-4);
-  width: 80%;
+  width: 100%;
 }
 
 .bodyEditor {
