+++ app/routes/privateRoutes/restClientPage/components/headers/Headers.module.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   gap: var(--spacing-6);
-  width: 80%;
+  width: 100%;
   align-items: center;
 }
