+++ app/routes/privateRoutes/restClientPage/components/urlBox/UrlBox.module.css
@@ -2,7 +2,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  width: 80%;
+  width: 100%;
   gap: var(--spacing-4);
 }
